pyproject.toml
[project]
name = "junefeed"
version = "0.1.0"
description = "A statefull RSS reader."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "feedparser>=6.0.11",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "textual>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["junefeed"]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
